fix: Remove externals file with gptel-namespaced macher variable (#27)
It wasn't really appropriate to add to the gptel namespace in an
external package. This is a breaking change if you were using presets to
specify a custom `macher-process-request-function`. If you need to
modify the `macher-process-request-function` as part of a gptel preset,
you can just create a variable alias manually:

```elisp
(defvaralias 'gptel--macher-process-request-function 'macher-process-request-function)
```

This change also prepares for a clean lint for melpa distribution.