fix: improve patch compatibility with `diff-mode` operations (#36)
Removes extra blank lines in patch metadata, which were causing issues
with `diff-apply-hunk` etc in some cases, for example when the new
content includes a "\ No trailing newline" marker.

Also adds a call to `diff-fixup-modifs` when rendering diffs, to
normalize descriptors of one-line changes to the format expected by
`diff-mode` ("@@ -1,1 +1,1 @@" instead of "@@ -1 +1 @@").

Note that if the patch contains a "\ No trailing newline" marker below
the _source_ changes (i.e. in the middle of a hunk), `diff-apply-hunk`
will still incorrectly prompt to "fix word-wrap damage". This appears to
be an underlying issue caused by a false positive from
`diff-sanity-check-hunk`.

Fixes #35.